+++ src/app/features/settings/Persona/ProfilesPage.tsx
@@ -13,7 +13,7 @@ export function PerMessageProfilePage({ requestClose }: PerMessageProfilePagePro
         <Box grow="Yes" gap="200">
           <Box grow="Yes" alignItems="Center" gap="200">
             <Text size="H3" truncate>
-              Profiles (Per-Message)
+              Persona
             </Text>
           </Box>
           <Box shrink="No">
